Rename some parts of the release checker
