+++ packages/cli/src/cliEntry.js
@@ -20,7 +20,7 @@ import assertRequiredOptions from './tools/assertRequiredOptions';
 import {logger} from '@react-native-community/cli-tools';
 import {setProjectDir} from './tools/packageManager';
 import findPlugins from './tools/findPlugins';
-import getNewerReleaseDataIfAvailable from './tools/getLatestRelease/getLatestRelease';
+import getLatestRelease from './tools/releaseChecker/getLatestRelease';
 import pkgJson from '../package.json';
 import loadConfig from './tools/config';
 
@@ -187,12 +187,12 @@ async function setupAndRun() {
 
   logger.setVerbose(commander.verbose);
 
-  const newerRelease = await getNewerReleaseDataIfAvailable();
-  if (newerRelease) {
+  const latestRelease = await getLatestRelease();
+  if (latestRelease) {
     logger.info('A newer version of React Native is available!');
     logger.info('');
-    logger.info(`Version: ${newerRelease.tag_name}`);
-    logger.info(`Changelog: ${newerRelease.html_url}`);
+    logger.info(`Version: ${latestRelease.tag_name}`);
+    logger.info(`Changelog: ${latestRelease.html_url}`);
     logger.info('');
     logger.info('Upgrade to the newer version by running react-native upgrade');
   }
